Tasks

+++ app/Http/Controllers/ProjectController.php
@@ -20,7 +20,7 @@ class ProjectController extends Controller
     public function index()
     {
     	
-    	return $this->repository->with(['owner','client'])->all();
+    	return $this->repository->with(['owner','client'])->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
     }
 
     public function store(Request $request)
@@ -30,18 +30,30 @@ class ProjectController extends Controller
 
     public function show($id)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
+
         return $this->service->find($id);
     }
 
     public function update(Request $request, $id)
     {
-    	return $this->service->update($request->all(), $id);
+    	if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
+
+        return $this->service->update($request->all(), $id);
 
     }
 
     public function destroy($id)
     {
-    	try {
+    	if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
+
+        try {
             $this->repository->find($id)->delete();
             return ['success'=>true, 'Projeto deletado com sucesso!'];
         } catch (QueryException $e) {
@@ -52,5 +64,28 @@ class ProjectController extends Controller
         } catch (\Exception $e) {
             return ['error'=>true, 'Ocorreu algum erro ao excluir o projeto.'];
         }
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($projectId, $userId);       
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->repository->hasMember($projectId, $userId);       
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) || $this->checkProjectMember($projectId)){
+            return true;
+        }
+
+        return false;
     }
 }
